# harmony-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harmony API",
    "description": "REST API for Harmony.\n\n**Rate Limiting:** All endpoints are subject to rate limiting. When the limit is exceeded, the API responds with `429 Too Many Requests` and a `Retry-After` header indicating when the client may retry.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/v1/channels/{id}/megolm-sessions": {
      "post": {
        "tags": [
          "Channels"
        ],
        "summary": "Register a Megolm session for an encrypted channel. Requires channel membership.",
        "description": "# Errors\nReturns `ApiError` if the channel is not encrypted, or on repository error.",
        "operationId": "create_megolm_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChannelId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMegolmSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Megolm session registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MegolmSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Channel not encrypted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/servers/{id}/channels": {
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "List all channels in a server.",
        "description": "# Errors\nReturns `ApiError` on repository error.",
        "operationId": "list_channels",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Channels"
        ],
        "summary": "Create a new channel in a server. Requires admin+ role.",
        "description": "# Errors\nReturns `ApiError` on validation failure or repository error.",
        "operationId": "create_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Channel created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/servers/{id}/channels/{channel_id}": {
      "delete": {
        "tags": [
          "Channels"
        ],
        "summary": "Delete a channel. Requires admin+ role.",
        "description": "# Errors\nReturns `ApiError` if this is the last channel or the channel is not found.",
        "operationId": "delete_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          },
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChannelId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Channel deleted"
          },
          "400": {
            "description": "Cannot delete last channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Channels"
        ],
        "summary": "Update a channel's name, topic, and/or flags. Requires admin+ role.\nEnabling encryption requires owner role (one-way toggle).",
        "description": "# Errors\nReturns `ApiError` on validation failure or repository error.",
        "operationId": "update_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          },
          {
            "name": "channel_id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChannelId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Channel updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a channel category.",
        "example": "aa0e8400-e29b-41d4-a716-446655440005"
      },
      "ChannelId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a channel within a server.",
        "example": "770e8400-e29b-41d4-a716-446655440002"
      },
      "ChannelListResponse": {
        "type": "object",
        "description": "Envelope for a list of channels (ADR-036).",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelResponse"
            }
          }
        }
      },
      "ChannelResponse": {
        "type": "object",
        "description": "Channel response returned to API consumers.",
        "required": [
          "id",
          "serverId",
          "name",
          "channelType",
          "position",
          "isPrivate",
          "isReadOnly",
          "encrypted",
          "slowModeSeconds",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "categoryId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CategoryId"
              }
            ]
          },
          "channelType": {
            "$ref": "#/components/schemas/ChannelType"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "encrypted": {
            "type": "boolean",
            "description": "Whether Megolm E2EE is enabled on this channel."
          },
          "id": {
            "$ref": "#/components/schemas/ChannelId"
          },
          "isPrivate": {
            "type": "boolean"
          },
          "isReadOnly": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "serverId": {
            "$ref": "#/components/schemas/ServerId"
          },
          "slowModeSeconds": {
            "type": "integer",
            "format": "int32",
            "description": "Minimum seconds between messages per user. 0 = disabled."
          },
          "topic": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChannelType": {
        "type": "string",
        "description": "Channel type (matches Postgres `channel_type` enum).",
        "enum": [
          "text",
          "voice"
        ]
      },
...
      "CreateChannelRequest": {
        "type": "object",
        "description": "Request body for creating a new channel.",
        "required": [
          "name"
        ],
        "properties": {
          "channelType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChannelType",
                "description": "Channel type (defaults to \"text\" if omitted)."
              }
            ]
          },
          "isPrivate": {
            "type": "boolean",
            "description": "Whether the channel is private (visible only to admin+ or explicit role grants)."
          },
          "isReadOnly": {
            "type": "boolean",
            "description": "Whether the channel is read-only (only admin+ can post messages)."
          },
          "name": {
            "type": "string",
            "description": "Channel name (lowercase alphanumeric + hyphens, 1-100 chars)."
          }
        },
        "additionalProperties": false
      },
...
      "CreateMegolmSessionRequest": {
        "type": "object",
        "description": "Request body for registering a Megolm session on a channel.",
        "required": [
          "sessionId"
        ],
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "The vodozemac Megolm session ID (base64-encoded Ed25519 public key)."
          }
        },
        "additionalProperties": false
      },
...
      "MegolmSessionId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a Megolm session record.",
        "example": "dd0e8400-e29b-41d4-a716-446655440008"
      },
      "MegolmSessionResponse": {
        "type": "object",
        "description": "Response after registering a Megolm session.",
        "required": [
          "id",
          "channelId",
          "sessionId",
          "createdAt"
        ],
        "properties": {
          "channelId": {
            "$ref": "#/components/schemas/ChannelId"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/MegolmSessionId",
            "description": "Server-generated record ID."
          },
          "sessionId": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 Problem Details response.\n\nAll API errors are returned in this standardized format.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "upgrade_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "WHY: Drives upsell when a plan limit is exceeded. The frontend renders\nan \"Upgrade\" CTA using this URL. Only present on `LimitExceeded` errors."
          }
        }
      },
...
      "ServerId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a server (guild).",
        "example": "660e8400-e29b-41d4-a716-446655440001"
      },
...
      "UpdateChannelRequest": {
        "type": "object",
        "description": "Request body for updating an existing channel.",
        "properties": {
          "encrypted": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Enable Megolm E2EE (one-way toggle: once true, cannot be set back to false)."
          },
          "isPrivate": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Update private flag (if provided)."
          },
          "isReadOnly": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Update read-only flag (if provided)."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "New channel name (if provided)."
          },
          "slowModeSeconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Minimum seconds between messages per user. 0 = disable slow mode. Admin+ only."
          },
          "topic": {
            "type": [
              "string",
              "null"
            ],
            "description": "New channel topic (if provided; null clears it)."
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT (mobile/API clients)"
      }
    }
  },
...
}